Teach dockerized-etc.zsh to optionally add a user
The dockerized-etc.zsh script now permits a -u|--user flag to set up a
new user for testing. This is useful for organic tetsing of scripts
which use sudo, and so any created user is also added as a sudoer will
ALL control.

Also be a bit more careful to HALFYAK_ prefix exported variable names.