Extract environment/command specific after files
Create a global halfyak_etc_after for the logic of "look for additional
configuration based on a file name", unifying the logic from zshenv's
platform specific code and zshrc's terminal specific code.

The common logic is in zshenv, which is a little ugly, but since it
needs it, there's not really a better place. The logic itself has some
semi duplication, but it's hard to factor it better in shell so i think
its ok for now.

The driver for the extraction is to use this elsewhere (notably in
platform specific setup).