Script to expand environment variables in a file
This is an alternative to envsubst (or similar) utilities which is
dependency free, which means i can run it as part of my setup.

Note that it is not pin-compatible with envsubst or shell expansion, the
expression form is more limited, there is no provision for escaping, and
most importantly non-existent variables are left in variable form.