Watch local mail spool and notify of mail
This uses hammerspoon to watch the local mail spool, and when mail is
received it places a message bubble in the toolbar, which can be clicked
to show mail Subject: headers.

This means i notice feedback from background scripts such as
LaunchAgents which are run using libexec/mail-on-fail to send mail if
they exit abnormally.

This contains a lightweight RFC822 parser in awk, because mail -H
truncates long subjects at a hard 80 column limit, which is too narrow.