Automatically select audio based on availablility
I have a specific fixed priority order for audio based on what is
available, and so i use Hammerspoon to select the devices based on this
when it sees changes in availability.

I'd like to extend this to things like automatically adjusting volume
depending on other context (such as location inferred from WiFi), but
i'll need a way to suppress changes for (for specific use cases). The
logic below is simple and addresses a majority of current cases.