Script to create throwaway volumes
The secure-ephemeral-volume script creates an encrypted filesystem with
a random discarded password and removes the image. With no arguments,
the volume can be used until unmounted. With arguments, they are
executed with cwd the ephemeral volume, and then it is unmounted.

Also add a weekly task to notify of stale attached disk images.

Fix an incorrect usage of trap in tests i discovered while making this
change.