A minimal vim color scheme
This mostly punts to a 16 color terminal. This works well for me using
iTerm, as i have it configured with a nice muted color scheme (solarized
dark). There's no need to get fancy with the colors here, as the
terminal is doing most of the work.

Also re-order my vimrc a little to make it work better when things go
wrong while editing the color files and interfering with vim loading.