Get zsh completion working for git "aliases"
This was surprisingly involved, because the Homebrew git provides a
completion which isn't built for this, but overrides the zsh provided
one which does have support for this use case. So i had to fiddle my
fpath (for macOS interactive only), and this meant moving the zshrc
after load a little later.

I'm not confident this is fully stable, but it works enough to take it
for a spin.