Option to dockerized-etc to mount cwd
Per help, dockerized-etc accepts -c or --cwd which mounts the host PWD
as /mnt/cwd.

Also set safe.directory for /mnt/etc.git inside docker. I'm not sure why
i started to need this - it's not only required for --cwd, and doesn't
even seem to be --user related. I'm wondering if a (default or user)
change to colima config showed this up, but a colima resintall didn't
make it unnecessary.