Script to convert schema.org/Recipe to ADF
This renders a simple confluence page from JSON+LD, such as that
extracted from a webpage with the recipe on it.

There's a current rough edge that that confluence takes the page title
in the envelope of the page update, not in the body, and ideally we'd
take the title from the JSON+LD. I've not addressed this properly yet, i
just manually set the title in the script which updated the page.