Disable automatic termguicolors inference
Recent iTerm2 is automatically detected as capable of 24 bit color,
which makes it ignore my color settings. Disable this for now - i'm not
exactly sure what changed in iTerm2. It ships its own terminfo for
reasons (via TERMINFO_DIRS) and there's an advanced setting to switch
that off, which might be worth exploring if other things have trouble
here.