Git hooks dispatcher and yolo blocker
A script to block pushing commits with the subject yolo, which is what
the git `yolo` alias creates. This is a defence against accidentally
pushing temporary commits.

Also include a dispatch script for invoking a directory of git hooks,
since i now have two hooks and so a single symlink is not sufficient.